feat: 도서 검색 관련 에러 타입 정의 및 적용과 확장함수 정의
