+++ android/TodokTodok/app/src/main/java/com/team/todoktodok/data/datasource/book/DefaultBookRemoteDataSource.kt
@@ -3,6 +3,7 @@ package com.team.todoktodok.data.datasource.book
 import com.team.domain.model.Book
 import com.team.todoktodok.data.network.request.CreateBookRequest
 import com.team.todoktodok.data.network.request.SaveBookRequest
+import com.team.todoktodok.data.network.request.toRequest
 import com.team.todoktodok.data.network.response.BookResponse
 import com.team.todoktodok.data.network.service.BookService
 import com.team.todoktodok.data.network.service.LibraryService
@@ -28,12 +29,7 @@ class DefaultBookRemoteDataSource(
     override suspend fun saveSelectedBook(book: Book): Long {
         val response =
             libraryService.createBook(
-                CreateBookRequest(
-                    bookIsbn = book.id,
-                    bookAuthor = book.author,
-                    bookImage = book.image,
-                    bookTitle = book.title,
-                ),
+                book.toRequest()
             )
         if (!response.isSuccessful) {
             throw IllegalStateException("서버 응답이 실패했습니다")
