+++ android/TodokTodok/app/src/main/java/com/team/todoktodok/data/network/request/CreateBookRequest.kt
@@ -1,5 +1,6 @@
 package com.team.todoktodok.data.network.request
 
+import com.team.domain.model.Book
 import kotlinx.serialization.Serializable
 
 @Serializable
@@ -8,4 +9,12 @@ data class CreateBookRequest(
     val bookTitle: String,
     val bookAuthor: String,
     val bookImage: String,
+)
+
+fun Book.toRequest() = CreateBookRequest(
+    bookIsbn = id,
+    bookTitle = title,
+    bookAuthor = author,
+    bookImage = image,
+
 )
