+++ android/TodokTodok/app/src/main/java/com/team/todoktodok/presentation/utview/discussion/SearchBookErrorType.kt
@@ -1,0 +1,8 @@
+package com.team.todoktodok.presentation.utview.discussion
+
+enum class SearchBookErrorType {
+    NO_SELECTED_BOOK,
+    BOOK_NOT_FOUND,
+    NO_SEARCH_RESULTS,
+    EMPTY_SEARCH_INPUT,
+}
