+++ android/TodokTodok/app/src/main/java/com/team/todoktodok/presentation/utview/discussion/vm/CreateDiscussionRoomViewModel.kt
@@ -116,9 +116,3 @@ class CreateDiscussionRoomViewModel(
     }
 }
 
-enum class SearchBookErrorType {
-    NO_SELECTED_BOOK,
-    BOOK_NOT_FOUND,
-    NO_SEARCH_RESULTS,
-    EMPTY_SEARCH_INPUT,
-}
